First cut at JIRA export dialog
This dialog plugs into the `:view_issues_show_details_bottom` hook in the
`issues/show` view. This is the most prominant place I have found to display
this information without hacking core Redmine/Rails code.

The box is styled using the `flash warning` CSS classes which are a core part
of Redmine themes. Possibly not a correct use of these classes, but good enough
for now.