Add YAML based plugin configuration
The plugin now loads settings from a local YAML file into the top-level module
object (which is hopefully a singleton). This is a hack. However, Redmine does
not provide an easy way to configure settings per-project and we need this
information to map Redmine project names to JIRA project names.

So. Loading local YAML doesn't feel right, but it is a "get shit done today" path
forward.