Handle export authorization in the controller
The first thing the Controller does is check if the user is authorized to
export issues. If they are not logged in, it will direct them to a login page.

This modification removes the authorization check from the view logic and adds
a clean redirection back to the issue view.

Fixes GH-12.